requirements.txt
fastapi>=0.104.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
supabase>=2.1.0
google-generativeai>=0.3.1
python-multipart>=0.0.6